one build file to rule them all
